Perform some restructuring
